First commit - loads of stuff being bundled in.
Currently we're at that point where loads of different subsystems need
working in simultaneously, as they all lean on each other.

The Emailer package is not implemented in mij as it depends on personal
credentials. It implements a simple render interface like everything else.

The login form currently renders a session key and sends it to the user's
email. The next step is to write an authentication module. This should
know how to render the navbar at the top right.

The next step will be for the login URL to actually set the cookie on the
user. Woo!