User can register, login and logout
When a user puts in their email we send them a URL with a sessionkey in
it. When clicked this sets a cookie in the user's browser with the key and
their email address. Ohohoh!

This should be done under the cover of SSL.

I am now going to work in submissions.