Critiques can be posted and that.
Critiques can be pushed on to the server. There are a variety of messages
for different kinds of failure.

It has been a while since I looked at this code properly. Commiting now so I
can fool around with making a Docker container for the Development version
of the platform.

After that it's a case of working in the rendering for Critiques, adding
the options to feature and flag posts and then doing the crude profile
pages. Might have to do something ti avatars too.

Gravatar, maybe? Who knows...?