Auto params for client, tatl and domain
For streamlining the params getting passed in to the modules we now auto
generate a handful. Gone through the modules and amended.

Changed 'user' to 'client' in a lot of places to make it less ambiguous.
That's because user was being used for the client in some places and for
the URI stuff too (which user the client is looking at).

Fucked off a bunch of modules that were no longer getting used.

Changed Keygen to be a module not a class.

If I get up to some more housekeeping I reckon some modules could be
folded in to each other. Thinking of changing 'render' to 'get' and
rolling the 'post' methods in to the same modules. I bet there's a way of
getting rid of a lot of the boilerplate crap.

Unrelated really - I had a go at bundling this up in Docker and it looks
promising.