Feature confirmation page worked in
When an editor goes to feature a page the confirmation screen pops up.
This is just to make sure the editor wants to feature the article.

If they click the 'hell no' button they are returned to the article.

Everything is rendered for the post request to work, it just needs wiring
in. This shouldn't take too long. I reckon the actual feature gunk will be
done in the user class.

This should come together nice and quick. =)

(I can see the next commit being in a few months time...)