Basic directory switching added for featuring.
Featured articles now get dumped in the featured directories. There are
chunks of code that are pretty ugly that need trimming down.

Refactoring...

Thinking of working in a class for Dynamic HTML Objects to make the code
an eensy bit more readable.

These changes are a little old (forgot about 'em) so commiting now for a
clean slate.