Critiques render on Article page
They look pretty swell. They render with the date and a dynamically
generated 'says' verb for fun. The username is also a hyperlink to their
profile.

Eventually I will work some kind of user thumbnail in. Should be quite
sexy. No rush for that bit though.

Just going to add the dates and buzz on to the Zine page.