Cosmetic changes, minor adjustments
Post cap added. User can now only put in five submissions.

A few useful additions to Dynamo macros. Really nothing special.
Here comes the profile page, once this is in and the auto delete is set
all we'll be needing is a few visual tweaks here and there and we should
be sorted.

Maybe I'll have it finished before the end of the year =D I'm hoping
before work on Monday - we'll see.