"Their" profile added - not finished.
Their profiles are now rendering with two tables of articles, one for
submissions and the other for featured. At some juncture I'd like to work
in a bio or something too.

When the user views their own profile they will be given the options to
delete their stuff from the platform.