Profile page for 'yours' state implemented
The delete button and the checkboxes appear next to the article titles in
the 'yours' state. This will allow the user to pick their own articles
that they want to delete.

When I work in this function it'll peel through the articles and choose
the checked ones to delete. Great stuff, huh?

The implementation was less complicated than I initially anticipated.
There may be artefacts pertaining to this dotted around the code, details
I no longer see...

Going live soon kittens, I can taste it!