RemoveArticle prompt worked in
The client gets an unordered list with their selected articles to delete
on. If the click the 'yeah' button the system will wind through the
client's articles and delete each one.

I made some minor changes to Tatl to make the meanings of the links a
little more obvious. Some of the changes may be reflected systemwide later
for instance Submissions may well be changed to Pending in other places...

I also made some amendments to the CSS and HTML to make sure other
resources are looked for in the root directory.