The seleted articles are now purged when deleted
If a client confirms the selection with the 'yeah' button on the
RemoveArticle page they are now stripped from the server.

I've made a couple of cosmetic changes too. The order of the buttons in
Tatl and the rephrasing submissions to pending... somewhere.

So close... not much else needs doing! =)