Register now shows an actual page
A prompt is show upon registration. Which is nice.

Made a couple of cosmetic changes and fixes.

If the user has now profile looking at a profile will force them to the
register page.

If they have no articles their profile page says so.

A weird issue when buzz is being counted on the Pending page was addressed
I was failing to escape the username so we were escaping the '+' sign.