Added additional images to the about page
I've tweaked the about page adding better images.

The first is one of me looking disdainful instead of the stupid one I
added previously. It's more... 'me'

The other one is kinda daft. It's 'bugs.png' and really it's two images.
It dynamically loads one on the fly - either me with code or me with
boobies. It's pretty cool. I'm pleased with it.

I also made a change to the setup script to check if '/mij' ecists on the
host before we try to build it. It' such a small change so it got worked
in with this. Woooo!