Added Twitter integration
The Twitterer sources are kept out of the main project for privacy reasons
It presents a simple interface "post" which just tweets the string we pass
in to it.

When I can be bothered updating the docker I will just write a dummy
module that dumps the string to stderr.

Nice, ey? Also made a small change to the about page.